+++ src/client/debugger/resources.tsx
@@ -424,12 +424,14 @@ export const GfxResourcesView: FunctionComponent<{
 		const copperScreens: { screen: IScreen; frames: number[] }[] = [];
 		// dupecheck copper screens from all frames
 		for(let i = 0; i < MODELS.length; i++) {
-			const copperScreen = GetScreenFromCopper(MODELS[i].copper, MODELS[0].amiga.chipsetFlags);
-			const dupe = copperScreens.findIndex((cs) => JSON.stringify(cs.screen) === JSON.stringify(copperScreen));
-			if(dupe === -1)
-				copperScreens.push({ screen: copperScreen, frames: [i + 1] });
-			else
-				copperScreens[dupe].frames.push(i + 1);
+			if(MODELS[i].copper) {
+				const copperScreen = GetScreenFromCopper(MODELS[i].copper, MODELS[0].amiga.chipsetFlags);
+				const dupe = copperScreens.findIndex((cs) => JSON.stringify(cs.screen) === JSON.stringify(copperScreen));
+				if(dupe === -1)
+					copperScreens.push({ screen: copperScreen, frames: [i + 1] });
+				else
+					copperScreens[dupe].frames.push(i + 1);
+			}
 		}
 
 		for(const cs of copperScreens) {